+++ fatiando/gravmag/__init__.py
@@ -10,7 +10,9 @@ magnetic field of various types of geometric objects:
 * :mod:`~fatiando.gravmag.prism`: 3D right rectangular prisms
 * :mod:`~fatiando.gravmag.polyprism`: 3D prisms with polygonal horizontal
   cross-sections
-* :mod:`~fatiando.gravmag.sphere`: Homogeneous spheres
+* :mod:`~fatiando.gravmag.sphere`: Spheres in Cartesian coordinates
+* :mod:`~fatiando.gravmag.tesseroid`: Tesseroids (spherical prisms) for modeling
+  in spherical coordinates
 * :mod:`~fatiando.gravmag.talwani`: 2D bodies with polygonal vertical
   cross-sections